finished project, testing overflow bit!

+++ 474Project4/Hardware.h
@@ -23,7 +23,7 @@ class Hardware
 public:
 	short int PC;
 	SAL** pcMemory;
-	std::map<std::string, int> symbolTable;
+	std::map<std::string, short> symbolTable;
 	short registerA_;
 	short registerB_;
 	int zeroResultBit_;
